feat: add use_jit config option for posenet
