--- peekingduck/nodes/model/posenetv1/posenet_files/decode_multi.py
@@ -1,251 +1,0 @@
-# Copyright 2018 Ross Wightman
-#
-# Modifications copyright 2022 AI Singapore
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      https://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""
-Supporting functions to decode multiple poses
-"""
-
-import operator
-from typing import List, Tuple, Union
-
-import numpy as np
-import scipy.ndimage as ndi
-import tensorflow as tf
-from numba import jit
-
-from peekingduck.nodes.model.posenetv1.posenet_files.constants import (
-    LOCAL_MAXIMUM_RADIUS,
-    MIN_ROOT_SCORE,
-    NMS_RADIUS,
-)
-from peekingduck.nodes.model.posenetv1.posenet_files.decode import decode_pose
-
-
-def decode_multiple_poses(
-    model_output: List[tf.Tensor],
-    pose_scores: np.ndarray,
-    poses: np.ndarray,
-    output_stride: int,
-    min_pose_score: float = 0.5,
-) -> int:
-    """Decodes the offsets and displacements map and return the keypoints
-
-    Args:
-        model_output (List[tf.Tensor]): Scores, offsets, displacements_fwd, and
-            displacements_bwd from model predictions.
-        pose_scores (np.array): Nx17 buffer to store keypoint scores where N is
-            the max persons to be detected.
-        poses (np.array): Nx17x2 buffer to store keypoints coordinate where N
-            is the max persons to be detected.
-        output_stride (int): Output stride to convert output indices to image
-            coordinates.
-        min_pose_score (float): Minimum pose score to return detected pose.
-
-    Returns:
-        pose_count (int): number of poses detected
-    """
-    scores = np.array(model_output[0][0])
-    offsets = np.array(model_output[1][0])
-    displacements_fwd = np.array(model_output[2][0])
-    displacements_bwd = np.array(model_output[3][0])
-
-    parts = _build_parts(scores, MIN_ROOT_SCORE)
-
-    offsets, displacements_fwd, displacements_bwd = _change_dimensions(
-        offsets, displacements_fwd, displacements_bwd
-    )
-
-    pose_count = _look_for_poses(
-        parts,
-        scores,
-        offsets,
-        displacements_fwd,
-        displacements_bwd,
-        pose_scores,
-        poses,
-        output_stride,
-        min_pose_score,
-    )
-
-    return pose_count
-
-
-def _build_parts(
-    scores: np.ndarray, score_threshold: float
-) -> List[Tuple[float, int, np.ndarray]]:
-    """Builds the parts list. Each part in the list consists of the score,
-    keypoint ID, and coordinates. A part is defined as the cell with the highest
-    score in the local window of size (D, D, 1) where
-    D = 2 * LOCAL_MAXIMUM_RADIUS + 1.
-
-    Args:
-        scores (np.ndarray): Score heatmap output by the model.
-        score_threshold (float): Local peaks below this threshold will be
-            discarded.
-
-    Returns:
-        (List[Tuple[float, int, np.ndarray]]): The parts list sorted by score in
-        descending order.
-    """
-    diameter = 2 * LOCAL_MAXIMUM_RADIUS + 1
-
-    local_peaks = ndi.maximum_filter(
-        scores, size=(diameter, diameter, 1), mode="constant"
-    )
-    parts = _build_parts_with_peaks(scores, score_threshold, local_peaks)
-    parts = sorted(parts, key=operator.itemgetter(0), reverse=True)
-    return parts
-
-
-@jit(nopython=True)
-def _build_parts_with_peaks(
-    scores: np.ndarray, score_threshold: float, local_peaks: np.ndarray
-) -> List[Tuple[float, int, np.ndarray]]:
-    """Builds the parts list. Each part in the list consists of the score,
-    keypoint ID, and coordinates. A part is defined as the cell with the highest
-    score in the local window of size (D, D, 1) where
-    D = 2 * LOCAL_MAXIMUM_RADIUS + 1.
-
-    Args:
-        scores (np.ndarray): Score heatmap output by the model.
-        score_threshold (float): Local peaks below this threshold will be
-            discarded.
-        local_peaks (np.ndarray): Local peaks in the score heatmap.
-
-    Returns:
-        (List[Tuple[float, int, np.ndarray]]): The parts list sorted by score in
-        descending order.
-    """
-    parts = []
-    local_peak_locations = (
-        (scores == local_peaks) & (scores > score_threshold)
-    ).nonzero()
-    for y_coord, x_coord, keypoint_id in zip(*local_peak_locations):
-        parts.append(
-            (
-                scores[y_coord, x_coord, keypoint_id],
-                keypoint_id,
-                np.array((y_coord, x_coord)),
-            )
-        )
-
-    return parts
-
-
-def _change_dimensions(*arrays: np.ndarray) -> Union[np.ndarray, List[np.ndarray]]:
-    """Changes the dimensions of input arrays from (h, w, x) to (h, w, x//2, 2)
-    to return a complete coordinate array.
-    """
-    results = []
-    for array in arrays:
-        height, width = array.shape[0], array.shape[1]
-        results.append(array.reshape(height, width, 2, -1).swapaxes(2, 3))
-
-    return results[0] if len(results) == 1 else results
-
-
-def _get_instance_score_fast(
-    existing_coords: np.ndarray,
-    nms_radius: int,
-    keypoint_scores: np.ndarray,
-    keypoints: np.ndarray,
-) -> float:
-    """Obtain instance scores for keypoints"""
-    if existing_coords.shape[0] > 0:
-        score_sum = np.linalg.norm(existing_coords - keypoints, axis=2) > nms_radius
-        not_overlapped_scores = np.sum(keypoint_scores[np.all(score_sum, axis=0)])
-    else:
-        not_overlapped_scores = np.sum(keypoint_scores)
-    return not_overlapped_scores / len(keypoint_scores)
-
-
-@jit(nopython=True)
-def _is_within_nms_radius(
-    existing_coords: np.ndarray, radius: int, point: np.ndarray
-) -> bool:
-    """Checks if the specified `point` is within the radius of any existing
-    coordinates.
-
-    Args:
-        existing_coords (np.ndarray): Array of existing coordinates.
-        radius (int): The radius/distance to check for.
-        point (np.ndarray): The specified point to check.
-    """
-    return existing_coords.shape[0] > 0 and np.any(
-        np.sum((existing_coords - point) ** 2, axis=1) <= radius
-    )
-
-
-def _look_for_poses(
-    parts: List[Tuple[float, int, np.ndarray]],
-    scores: np.ndarray,
-    offsets: np.ndarray,
-    displacements_fwd: np.ndarray,
-    displacements_bwd: np.ndarray,
-    pose_scores: np.ndarray,
-    poses: np.ndarray,
-    output_stride: int,
-    min_pose_score: float,
-) -> int:
-    # pylint: disable=too-many-arguments, too-many-locals
-    """Change dimensions from (h, w, x) to (h, w, x//2, 2) to return a
-    complete coordinate array
-    """
-    pose_count = 0
-    max_pose_detections = pose_scores.shape[0]
-    squared_nms_radius = NMS_RADIUS ** 2
-
-    for root_score, root_id, root_coords in parts:
-        root_image_coords = (
-            root_coords * output_stride
-            + offsets[root_coords[0], root_coords[1], root_id]
-        )
-
-        if _is_within_nms_radius(
-            poses[:pose_count, root_id], squared_nms_radius, root_image_coords
-        ):
-            continue
-
-        keypoint_scores = pose_scores[pose_count]
-        keypoints = poses[pose_count]
-        decode_pose(
-            root_score,
-            root_id,
-            root_image_coords,
-            scores,
-            offsets,
-            output_stride,
-            displacements_fwd,
-            displacements_bwd,
-            keypoint_scores,
-            keypoints,
-        )
-
-        pose_score = _get_instance_score_fast(
-            poses[:pose_count],
-            NMS_RADIUS,
-            keypoint_scores,
-            keypoints,
-        )
-        if pose_score >= min_pose_score:
-            pose_count += 1
-        else:
-            pose_scores[pose_count] = 0
-
-        if pose_count >= max_pose_detections:
-            break
-
-    return pose_count
